Update sexual_harassment_policy.md
I wanted to add a statement of purpose for the harassment policy. I believe it is a good idea to make this policy mandatory for every collaborator to read and agree to abide by to ensure a more welcoming workplace. By stating the purpose of HERA wanting to work towards becoming anti-inequality, collaborators understand that agreeing to this policy is just one part of the bigger picture that HERA is trying to become more inclusive to everyone.